refactor search.h
