+++ src/move_search/types.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "move_generation/types.h"
+#include "search_constants.h"
 
 struct ScoredMove {
 	Move move;
@@ -7,3 +8,29 @@ struct ScoredMove {
 };
 
 typedef std::vector<ScoredMove> ScoredMoves;
+
+struct BestMoveSearchParameters {
+	short depth = MAX_DEPTH;
+	int hard_time_limit = DEFAULT_SEARCH_TIME;
+	int soft_time_limit = DEFAULT_SEARCH_TIME;
+	bool debug_info = false;
+};
+
+typedef Move Line[MAX_DEPTH];
+extern std::ostream& operator<<(std::ostream& os, const Line& line);
+
+struct BestMoveSearchResults {
+	Move best_move = Move();
+	Line pv;
+	int depth_searched = 0;
+	int seldepth = 0;
+
+	int time_searched = 0;
+	int value = 0;
+
+	uint64_t q_nodes_searched = 0;
+	uint64_t nodes_searched = 0;
+	double nodes_per_second = 0;
+};
+
+extern std::ostream& operator<<(std::ostream& os, const BestMoveSearchResults& results);
